Adding 'Edit with Notepad++' file association option
The '->' button is changed to a 'Open ->' and 'Edit ->' button. Edit command associations are colored blue.

Fix: Adding file associations to existing file type (rather than changing the file type to Notepad++_file) so that the file icon is retained

(Removing a file association restores the backup value as before)